Changes to all files.
I added some more functionality to the flappy bird game so far. The game now detects when the bird hits the side of a pipe, but it does it know if the bird hits the tops of the pipes. Also, I added some memory management, i.e. pipes are deleted when they are not present on the screen.