Kill states
This commit adds a death state to the bird (comes with a color change).  I created conditionals to kill if above or below the window, and if within a pipe.  I also subtly changed the pipe create code to allow us to start wider and gradually narrow the gap as a player gets better.  I'll save that for next update.