Lots of changes!
I added pipe sprites to help complete the look of the game. I also changed the spacing and speed of the pipes so that the game is a consistent level of difficulty. I didnt want it to continually get difficult to a point. I wanted the game to have a fixed level of difficulty. I also added a score counter to asses how many pipes the player has passed.